+++ mycodo/mycodo_flask/camera/base_camera.py
@@ -53,19 +53,25 @@ class CameraEvent(object):
 
 
 class BaseCamera(object):
-    thread = None  # background thread that reads frames from camera
-    frame = None  # current frame is stored here by background thread
-    last_access = 0  # time of last client access to the camera
-    event = CameraEvent()
+    thread = {}  # background thread that reads frames from camera
+    frame = {}  # current frame is stored here by background thread
+    last_access = {}  # time of last client access to the camera
+    event = {}
 
-    def __init__(self, opencv_device=None):
+    def __init__(self, camera_type=None, device=None):
         """Start the background camera thread if it isn't running yet."""
-        if BaseCamera.thread is None:
-            BaseCamera.last_access = time.time()
+        self.unique_name = "{cam}_{dev}".format(cam=camera_type, dev=device)
+        print("TEST01: {}".format(self.unique_name))
+        BaseCamera.event[self.unique_name] = CameraEvent()
+        if self.unique_name not in BaseCamera.thread:
+            BaseCamera.thread[self.unique_name] = None
+        if BaseCamera.thread[self.unique_name] is None:
+            BaseCamera.last_access[self.unique_name] = time.time()
 
             # start background frame thread
-            BaseCamera.thread = threading.Thread(target=self._thread, args=(opencv_device,))
-            BaseCamera.thread.start()
+            BaseCamera.thread[self.unique_name] = threading.Thread(target=self._thread,
+                                                                   args=(self.unique_name,))
+            BaseCamera.thread[self.unique_name].start()
 
             # wait until frames are available
             while self.get_frame() is None:
@@ -73,36 +79,33 @@ class BaseCamera(object):
 
     def get_frame(self):
         """Return the current camera frame."""
-        BaseCamera.last_access = time.time()
+        BaseCamera.last_access[self.unique_name] = time.time()
 
         # wait for a signal from the camera thread
-        BaseCamera.event.wait()
-        BaseCamera.event.clear()
+        BaseCamera.event[self.unique_name].wait()
+        BaseCamera.event[self.unique_name].clear()
 
-        return BaseCamera.frame
+        return BaseCamera.frame[self.unique_name]
 
     @staticmethod
-    def frames(opencv_device):
+    def frames():
         """"Generator that returns frames from the camera."""
         raise RuntimeError('Must be implemented by subclasses')
 
     @classmethod
-    def _thread(cls, opencv_device):
+    def _thread(cls, unique_name):
         """Camera background thread."""
         print('Starting camera thread')
-        if opencv_device is not None:
-            print('opencv device number {dev} selected'.format(
-                dev=opencv_device))
-        frames_iterator = cls.frames(opencv_device)
+        frames_iterator = cls.frames()
         for frame in frames_iterator:
-            BaseCamera.frame = frame
-            BaseCamera.event.set()  # send signal to clients
+            BaseCamera.frame[unique_name] = frame
+            BaseCamera.event[unique_name].set()  # send signal to clients
             time.sleep(0)
 
             # if there hasn't been any clients asking for frames in
-            # the last 10 seconds then stop the thread
-            if time.time() - BaseCamera.last_access > 10:
+            # the last 5 seconds then stop the thread
+            if time.time() - BaseCamera.last_access[unique_name] > 5:
                 frames_iterator.close()
                 print('Stopping camera thread due to inactivity')
                 break
-        BaseCamera.thread = None
+        BaseCamera.thread[unique_name] = None